feat: add devcontainer features to setup dev env
Some commands and ports are enabled and configured for a development
environment. Configured postCerateCommand installs dependencies so that
it doesn't need to be installed manually every time devcontainer is
run. postStartCommand starts the app when the container is run. The
ports 3306 and 8000 are exposed on the dev container setup for
accessing the application and supporting mysql database.