Update root path to /app (#3)
* feat: add devcontainer features to setup dev env

Some commands and ports are enabled and configured for a development
environment. Configured postCerateCommand installs dependencies so that
it doesn't need to be installed manually every time devcontainer is
run. postStartCommand starts the app when the container is run. The
ports 3306 and 8000 are exposed on the dev container setup for
accessing the application and supporting mysql database.

* feat: add jwt authentication and token validation

Added jwt authetication on some endpoints for testing and implemented
jwt token validation based on jwks uri. Token validation aids in the
authenticity of the token and protects against token forgery.

* feat: add prometheus instrumentation

Added prometheus instrumetation to collect metrics from the application
and visualize it using grafana.

* Update devcontainer.json

* Update Dockerfile

* Update Dockerfile

* Update devcontainer.json

* fix: dev container not working in codespace

* Update docker-compose.yml

* Update docker-compose.yml

* Update devcontainer.json

* Update devcontainer.json

* Update devcontainer.json

* Update devcontainer.json

* Update docker-compose.yml

* Update docker-compose.yml

* feat: update root path url

The root path url has been set to /app to facilitate the hosting using
nginx path based routing.