CI related changes merged to main (#5)
* feat: add devcontainer features to setup dev env

Some commands and ports are enabled and configured for a development
environment. Configured postCerateCommand installs dependencies so that
it doesn't need to be installed manually every time devcontainer is
run. postStartCommand starts the app when the container is run. The
ports 3306 and 8000 are exposed on the dev container setup for
accessing the application and supporting mysql database.

* feat: add jwt authentication and token validation

Added jwt authetication on some endpoints for testing and implemented
jwt token validation based on jwks uri. Token validation aids in the
authenticity of the token and protects against token forgery.

* feat: add prometheus instrumentation

Added prometheus instrumetation to collect metrics from the application
and visualize it using grafana.

* Update devcontainer.json

* Update Dockerfile

* Update Dockerfile

* Update devcontainer.json

* fix: dev container not working in codespace

* Update docker-compose.yml

* Update docker-compose.yml

* Update devcontainer.json

* Update devcontainer.json

* Update devcontainer.json

* Update devcontainer.json

* Update docker-compose.yml

* Update docker-compose.yml

* feat: update root path url

The root path url has been set to /app to facilitate the hosting using
nginx path based routing.

* Feature/implement GitHub actions (#4)

* feat: implement ci actions

* feat: run linters in ci

* feat: run unit-tests in ci

* feat: docker-build image in ci

* feat: build and push image to dockerhub

* feat: implement image tagging in docker

* feat: implement image tagging in docker

* feat: implement image tagging in docker

* feat: run workflows on tags

* feat: run workflows on tags

* feat: disable docker build record

* fix: correct Docker image tagging patterns and update build summary environment variable

* fix: remove extraneous character from Docker image tagging pattern

* fix: update Docker image tagging pattern for SHA to include short format

* feat: implement image update on deployment repo

* feat: implement image tag update on deployment repo

* fix: updated build to run on tags

* fix: updated build to run on tags

* fix: updated build to run on tags

* fix: updated to replace image name and tag

* fix: updated to replace image name and tag

* fix: update README.md

* fix: correct flake8 linting errors

It corrects linting errors reported by flake8 to ensure consistency in
formatting the code.

* fix: correct flake8 reported end of line

flake8 reported end of line linting issue. It has been resolved.